Logbook userpoint icon now displayed offset from airport icon. Hotspot of logbook userpoint is now top left of the symbol. #366

+++ src/mapgui/mappainterweather.cpp
@@ -90,9 +90,9 @@ void MapPainterWeather::render(PaintContext *context)
 void MapPainterWeather::drawAirportWeather(PaintContext *context,
                                            const atools::fs::weather::Metar& metar, float x, float y)
 {
-  int size = context->sz(context->symbolSizeAirport, context->mapLayerEffective->getAirportSymbolSize());
+  float size = context->sz(context->symbolSizeAirport, context->mapLayerEffective->getAirportSymbolSize());
   bool windBarbs = context->mapLayer->isAirportWeatherDetails();
 
-  symbolPainter->drawAirportWeather(context->painter, metar, x - size * 4 / 5, y - size * 4 / 5, size,
+  symbolPainter->drawAirportWeather(context->painter, metar, x - size * 4.f / 5.f, y - size * 4.f / 5.f, size,
                                     true /* Wind pointer*/, windBarbs, context->drawFast);
 }
